src/fin_chat_agent/requirements.txt
fastapi==0.115.12
uvicorn
google-adk>=0.1.0
google-cloud-storage>=2.10.0
python-dotenv==1.1.0
pydantic==2.11.7
pdfplumber>=0.10.0
google-generativeai>=0.5.4
google-cloud-aiplatform>=1.38.0
